Remove TODO about &rest syntax for cmds
On second thought, I don't think this is that critical.
It might make sense to have it, but commands aren't
as general as functions, and I can't immediately
think of a need for it.

It could also be implemented as an argtype if someone
really wanted it (but really didn't want to contribute
to rainboots core, for whatever reason)