Fail expand-args if there is extra, unparsed input
This is (probably) important for multi-arity cmds,
but may not be a complete solution. For example, if the
given command *only has* a single arity, it may be safe
to assume any input is intended for that command, even
if there is extra input.

We may also want to support the [& rest] syntax in
regular clojure fns, especially with argtypes. May not
be a priority, but I can certainly imagine scenarios
when you'd like to accept a variable number of
similarly-typed input. The [& {:keys []}] syntax does
not make a whole lot of sense, however....