Use vectors for hiccup-like syntax in `send!` (#7)
* Initial refactor to only treat non-vector sequences as varargs

Refs #6

* First-pass implementation of hiccup parsing

Refs #6

This may not be the most performant solution, but at least it should
detect invalid input. Includes coloring using the builtin colors.

* Declare `*handlers*` as `^:dynamic` upfront

* Add prewalk import

* Add defhandler for registering new keywords

Refs #6

This change also guarantees the `:rainboots.hiccup/string` core handler,
which shouldn't need to be used directly.

* Refactor so children are varags (like normal hiccup); simplify spec

This is much easier to reason about, and clearer on the fn param
declaration part. Nice!

* Add doc param to defhandler

* Properly support "any" argument to a hiccup handler

* Avoid re-processing simple string sequences

* Describe hiccup usage in the README