Begin experiments with hot-restart support
The traditional technique is to write the file descriptors to disk,
exec a new process in a way that shares file descriptors, then read
the file descriptors off the disk in the new process and reconnect
those sockets with the account data. This is tricky in Java since
we don't have *real*, *reliable* access to the FD, but this commit
includes some hacks.

The next magic trick will be reconstructing the manifold/aleph stream
from the file descriptor. It should be possible to just instantiate a
`FileDescriptor` with reflection, then perhaps we can wrap that in a
`FileInputStream` and get a `FileChannel` from that. From there we can
use aleph's `source` and `sink` methods with the channel, and `splice`
them together using Manifold.